auth has been all set up

+++ graphql/resolvers/booking.js
@@ -3,41 +3,50 @@ const Booking = require('../../models/booking');
 const { dateToString } = require('../../helpers/date');
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     try {
       const bookings = await Booking.find();
-      return bookings.map(booking => {
+      return bookings.map((booking) => {
         return {
           ...booking._doc,
           createdAt: dateToString(booking._doc.createdAt),
-          updatedAt: dateToString(booking._doc.updatedAt)
+          updatedAt: dateToString(booking._doc.updatedAt),
         };
-      })
+      });
     } catch (err) {
       throw err;
     }
   },
-  bookEvent: async (args) => {
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
     const fetchedEvent = await Event.findOne({ _id: args.eventId });
     const booking = new Booking({
-      user: '5f0fa27bf05326694cdd33c6',
-      event: fetchedEvent
+      user: req.userId,
+      event: fetchedEvent,
     });
     const result = await booking.save();
     return {
       ...result._doc,
       createdAt: dateToString(result._doc.createdAt),
-      updatedAt: dateToString(result._doc.updatedAt)
+      updatedAt: dateToString(result._doc.updatedAt),
+    };
+  },
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
+    try {
+      const booking = await Booking.findById(args.bookingId);
+      const event = { ...booking.event._doc, _id: booking.event.id };
+      await Booking.deleteOne({ _id: args.bookingId });
+      return event;
+    } catch (err) {
+      throw err;
     }
   },
-  cancelBooking: async (args) => {
-     try{
-       const booking = await Booking.findById(args.bookingId);
-       const event = {...booking.event._doc, _id: booking.event.id}
-       await Booking.deleteOne({_id: args.bookingId})
-       return event;
-     } catch (err){
-       throw err;
-     }
-  }
 };
